// Anova.Evolve.Admin.Frontend/frontend/src/containers/DataChannelEditor/components/ProfileTab/components/IntegrationProfileTableDrawer/types.ts
export interface Values {
  integrationProfile1: {
    isIntegrationEnabled: boolean;
    integrationId: string | null;
    integrationDomainId: string | null;
  };
  integrationProfile2: {
    isIntegrationEnabled: boolean;
    integrationId: string | null;
    integrationDomainId: string | null;
  };
}